+++ packages/orq-rc/src/models/operations/retrievedatasource.ts
@@ -172,7 +172,7 @@ export const RetrieveDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01K7KZ4WXDBY9DNH03W6P5B62C"),
+  _id: z.string().default("01K7M9PT2EFST1VB54ZM30ZDPG"),
   display_name: z.string(),
   description: z.string().optional(),
   status: RetrieveDatasourceStatus$inboundSchema,
@@ -216,7 +216,7 @@ export const RetrieveDatasourceResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   RetrieveDatasourceResponseBody
 > = z.object({
-  id: z.string().default("01K7KZ4WXDBY9DNH03W6P5B62C"),
+  id: z.string().default("01K7M9PT2EFST1VB54ZM30ZDPG"),
   displayName: z.string(),
   description: z.string().optional(),
   status: RetrieveDatasourceStatus$outboundSchema,