+++ packages/orq-rc/src/models/operations/updatedatasource.ts
@@ -244,7 +244,7 @@ export const UpdateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01K7KZ4WXEDV7QDHGP59F7V572"),
+  _id: z.string().default("01K7M9PT2FYJG0CQ1DCX97EY7A"),
   display_name: z.string(),
   description: z.string().optional(),
   status: UpdateDatasourceStatus$inboundSchema,
@@ -288,7 +288,7 @@ export const UpdateDatasourceResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   UpdateDatasourceResponseBody
 > = z.object({
-  id: z.string().default("01K7KZ4WXEDV7QDHGP59F7V572"),
+  id: z.string().default("01K7M9PT2FYJG0CQ1DCX97EY7A"),
   displayName: z.string(),
   description: z.string().optional(),
   status: UpdateDatasourceStatus$outboundSchema,