+++ packages/orq-rc/RELEASES.md
@@ -4289,3 +4289,13 @@ Based on:
 - [typescript v3.14.0-rc.29] packages/orq-rc
 ### Releases
 - [NPM v3.14.0-rc.29] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.29 - packages/orq-rc
+
+## 2025-10-15 16:03:42
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.636.3 (2.723.11) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.14.0-rc.30] packages/orq-rc
+### Releases
+- [NPM v3.14.0-rc.30] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.30 - packages/orq-rc